+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v2.0.4
 
+新增: trade_pro.go ReqQryTradingAccount ReqQryPositionDetail ReqQryPosition 函数
 新增: 出入金函数
 更新: String 函数只取 \x00 前的数据
 升级: CTP v6.6.9